+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+-State Bundle 1.2.0 (2017-09-27)
+--------------------------------
+- Added method `copyToBundle(Bundle)` which can copy the contents of a StateBundle into an android.os.Bundle.
+
 -State Bundle 1.1.5 (2017-05-14)
 --------------------------------
 - Better error message when `bundle.putAll()` is called with `null` (invalid) argument.